Use API wrapper to simplify fetch logic
